NVM
These commits add NVM support. The only addition to the API are the
dismount and mount functions

This is a squash of 11 commits:

Some cleanup

added some rudimentary benchmarking stuff

minor script stuff

cleaning up

minor

random slot choice on insert

fixed a minor bug

removed deprecated benchmark.cc

remove restrict

durable level 3

NVM: Added Dismount and Mount

Added a dismount function, which is really just a deinit, but which
leaves behind the files.

Added a mount function which reads the files and recreates the contents
of the hash table.

Added to main in between inserts and queries.